+++ packages/stash/lib/src/api/cache/cache_manager.dart
@@ -31,7 +31,7 @@ abstract class CacheManager {
   /// * [sampler]: The sampler to use upon eviction of a cache element, defaults to [FullSampler] if not provided
   /// * [evictionPolicy]: The eviction policy to use, defaults to [LruEvictionPolicy] if not provided
   /// * [maxEntries]: The max number of entries this cache can hold if provided. To trigger the eviction policy this value should be provided
-  /// * [cacheLoader]: The [CacheLoader] that should be used to fetch a new value upon expiration
+  /// * [cacheLoader]: The [CacheLoader] that should be used to fetch a new value upon absence expiration
   /// * [clock]: The source of time to be used on this, defaults to the system clock if not provided
   /// * [eventListenerMode]: The event listener mode of this cache
   /// * [statsEnabled]: If statistics should be collected, defaults to false